Update install and fixed polling problem
.gitignore - A bunch of files it would be good not to check in :)

test-setup.sh - To adapt to the new way of building PouchDB I now clone
PouchDB into its own separate directory, pouchdbclone and then do an
install and build, symbolically link to packages/pouchdb and install that
as well.

test-pouchdb-minimum.sh - The tests (as opposed to the live code) now
lives in pouchdbclone

test-pouchdb.sh - Same as previous

changes - As explained in #326 I tried to hook all the obvious events and
literally nothing fired when a connection was lost. So instead when
the heartbeat runs I check if we have lost the socket and if so then
I clean up. I also made cleanup able to cancel any live changes objects
and call res.end.